+++ terraform/main.tf
@@ -239,6 +239,7 @@ module "cne_instance" {
   cneinstance_hugepages_profile_name = var.cneinstance_hugepages_profile_name
 
   cneinstance_tmm_replicas                   = var.cneinstance_tmm_replicas
+  cneinstance_storage_class_name             = var.cneinstance_storage_class_name
   cneinstance_watch_namespaces               = var.cneinstance_watch_namespaces
   cneinstance_tmm_anti_affinity              = var.cneinstance_tmm_anti_affinity
   cneinstance_tmm_anti_affinity_topology_key = var.cneinstance_tmm_anti_affinity_topology_key